Refactor secret manager
